Revert "refactor: 상품 패키지 -> 메뉴 패키지로 이동"
This reverts commit c2a443bf01d8b29b58b207745e212de0fce84e3b.
