+++ src/test/java/kitchenpos/common/fixtrue/ProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.common.fixtrue;
 
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
